refactor: update register for new backend
