+++ api/models/user.ts
@@ -1,23 +1,18 @@
 export interface UserPersonalData {
-    id: string,
-    contact_phone?: string,
+    id: number,
+    patient_id: number
+    phone?: string,
     weight?: string,
     height?: string,
-    other_info?: string,
-    created_at: string,
-    updated_at: string,
-    deleted_at?: string
+    info?: string,
 }
 
 export interface User {
-    id: string,
+    id: number,
     name: string,
     email: string,
-    email_verified_at: string,
-    type_id: 'DOCTOR' | 'PATIENT' | 'ADMIN',
-    personal_data: UserPersonalData,
-    created_at: string,
-    updated_at: string,
-    deleted_at?: string
-
+    roles: {
+        name: 'doctor' | 'patient'
+    },
+    info: UserPersonalData
 }
